Customer Summary Report Finished with test
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C23769

+++ TOneV2/Code/TOneV2/TOne.Analytics.Entities/Billing/CustomerServices.cs
@@ -9,6 +9,6 @@ namespace TOne.Analytics.Entities
     public class CustomerServices
     {
         public string AccountId { get; set; }
-        public decimal Services { get; set; }
+        public double Services { get; set; }
     }
 }
